Exercício 024
Implemente um controle simples de mercadorias em uma despensa doméstica. Para cada produto armazene um código numérico, descrição e quantidade atual. O programa deve ter opções para entrada e retirada de produtos, bem como um relatório geral e um de produtos não disponíveis. Armazene os dados em arquivo binário.

+++ ExerciciosResolvidos/ex024.c
@@ -40,7 +40,7 @@ void retiraProduto(FILE *arquivo, int codigo){
 			//Atualiza a quantidade do produto no arquivo
 			produto.quantidade--;
 
-			fseek(arquivo, -sizeof(struct Produto), SEEK_CUR); //Move o ponteiro de volta para a posicao correta
+			fseek(arquivo, -((long long int)sizeof(struct Produto)), SEEK_CUR); //Move o ponteiro de volta para a posicao correta
 			fwrite(&produto, sizeof(struct Produto), 1, arquivo);
 
 			printf("Produto retirado com sucesso!\n");
